backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
google-genai>=1.0.0
numpy==1.26.4
youtube-transcript-api==0.6.2
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
pymongo==4.6.1
